Update Forge lock 2
This is the updated forge lock 2, I fixed an unnecessary while loop as well as some added some features.
------------------------------------------------------------------------
Operation is like this:
Once you press the button and the cabinet has not been opened you have 4 seconds to scan an ID. If the scan fails or you did not scan (meaning the cabinet remains closed), you must press the button again. If the scan works (meaning the cabinet successfully opens) and the cabinet is currently closed (please see "closed sensor" below for more info), if the button is pressed again, it will lock the cabinet, or if the cabinet is closed for more than 15 seconds, it will automatically close. Both the button lock as well as the automatic close will start the sleeping sequence. 
-----------------------------------------------------------------------
For the wifi. What we need to do is in the function "getScanResults" we need to connect to the wifi network and the server, and instead of "digitalRead(RFID)" located in lines 218 and 221, we need to connect to the IEEE server and return whether the user was accepted or not. I will need help with this.
-----------------------------------------------------------------------
For the sleep sequence we simply need to insert it. We might come up with some problems which we will have to deal with as I am unsure if this code will run as smoothly as I hope after wake up.
----------------------------------------------------------------------
Closed Sensor - For the closed sensor I was thinking of using a reed switch instead of a light sensor as it would allow for operation of the cabinet lock even under poor lighting conditions, but a light sensor will work as well.